Title: 시력 검사 2, Time: 107ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240724/시력 검사 2/eye-test-2.py
@@ -1,4 +1,4 @@
-score = int(input())
+score = float(input())
 
 if score >= 1.0:
     print("High")
